Move DaprClient into init file and delete main.py.
